refactor: split dashboard subcomponents
