+++ components/report.tsx
@@ -47,33 +47,16 @@ import {
   Shield
 } from "lucide-react";
 import { DropdownPDFExport } from "./pdf-export-wrapper";
-
-interface StatKV {
-  key: string;
-  value: number;
-}
+import MaturityCard from "./report/MaturityCard";
+import ROICard from "./report/ROICard";
+import OpportunityCard from "./report/OpportunityCard";
+import SummaryCard from "./report/SummaryCard";
+import TasksByAreaSection from "./report/TasksByAreaSection";
+import type { StatKV, MaturityData, ROIData } from "./report/types";
 
 interface Props {
   companyName: string;
   stats: StatKV[];
-}
-
-interface MaturityData {
-  score: number;
-  level: string;
-  color: string;
-  description: string;
-}
-
-interface ROIData {
-  current: number;
-  potential: number;
-  opportunity: number;
-  employeeCount: number;
-  currentHoursPerWeek: number;
-  currentMinutesSavedPerDay: number;
-  adoptionEffectiveness: number;
-  missedHoursPerDay?: number;
 }
 
 const COLORS = [
@@ -169,342 +152,6 @@ const getROIFromStats = (stats: StatKV[]) => {
     adoptionEffectiveness: singleValue(stats, "roi.adoptionEffectiveness", 1),
     missedHoursPerDay: singleValue(stats, "roi.missedHoursPerDay", 2)
   };
-};
-
-const MaturityCard: React.FC<{ maturity: MaturityData }> = ({ maturity }) => (
-  <Card className="border-2" style={{ borderColor: maturity.color }}>
-    <CardHeader className="pb-2">
-      <CardTitle className="text-sm font-medium text-muted-foreground flex items-center gap-2">
-        <Target className="w-4 h-4" />
-        Nivel de Madurez IA
-      </CardTitle>
-    </CardHeader>
-    <CardContent className="pt-0">
-      <div className="flex items-center gap-4">
-        <div className="text-4xl font-bold" style={{ color: maturity.color }}>
-          {maturity.score}
-        </div>
-        <div>
-          <Badge style={{ backgroundColor: maturity.color, color: 'white' }}>
-            {maturity.level}
-          </Badge>
-          <p className="text-sm text-muted-foreground mt-1">
-            {maturity.description}
-          </p>
-        </div>
-      </div>
-      <div className="mt-3 bg-muted/40 rounded-full h-3">
-        <div
-          className="h-3 rounded-full transition-all duration-500"
-          style={{ 
-            width: `${maturity.score}%`, 
-            backgroundColor: maturity.color 
-          }}
-        />
-      </div>
-    </CardContent>
-  </Card>
-);
-
-const ROICard: React.FC<{ roi: ROIData }> = ({ roi }) => {
-  // roi.current, roi.potential, and roi.opportunity are now TOTAL MONTHLY values.
-  // No need to divide by 12 anymore.
-  const currentMonthlyTotal = roi.current;
-  const potentialMonthlyTotal = roi.potential;
-  const opportunityMonthlyTotal = roi.opportunity;
-
-  return (
-    <Card className="border-2 border-orange-200 bg-orange-50/50">
-      <CardHeader className="pb-2">
-        <CardTitle className="text-sm font-medium text-muted-foreground flex items-center gap-2">
-          <DollarSign className="w-4 h-4" />
-          Análisis ROI Mensual ({roi.employeeCount} empleados)
-        </CardTitle>
-      </CardHeader>
-      <CardContent className="pt-0 space-y-3">
-        {/* Current Value */}
-        <div className="flex justify-between items-center p-2 bg-blue-50 rounded border border-blue-200">
-          <span className="text-sm font-medium text-blue-700">💰 Valor actual generado (total):</span>
-          <span className="font-bold text-blue-700">${currentMonthlyTotal.toLocaleString()}/mes</span>
-        </div>
-        
-        {/* Opportunity Cost - What they're losing */}
-        <div className="flex justify-between items-center p-2 bg-red-50 rounded border border-red-200">
-          <span className="text-sm font-medium text-red-700">⚠️ Dinero que pierden (total):</span>
-          <span className="font-bold text-red-700">${opportunityMonthlyTotal.toLocaleString()}/mes</span>
-        </div>
-        
-        {/* Maximum Potential */}
-        <div className="flex justify-between items-center p-2 bg-green-50 rounded border border-green-200">
-          <span className="text-sm font-medium text-green-700">🎯 Potencial con training (total):</span>
-          <span className="font-bold text-green-700">${potentialMonthlyTotal.toLocaleString()}/mes</span>
-        </div>
-        
-        {/* Context */}
-        <div className="border-t pt-2 text-xs text-muted-foreground space-y-1">
-          <div><strong>Tiempo ahorrado (reportado):</strong> {roi.currentMinutesSavedPerDay.toFixed(1)} min/día por empleado.</div>
-          <div><strong>Eficiencia de Adopción IA (real):</strong> {roi.adoptionEffectiveness.toFixed(1)}%.</div>
-          
-          {/* Display missed hours if not 100% effective and missedHoursPerDay is available and numeric */}
-          {roi.adoptionEffectiveness < 99.9 && typeof roi.missedHoursPerDay === 'number' && (
-            <div className="text-gray-600">
-              ↳ Potencial no alcanzado: {roi.missedHoursPerDay.toFixed(2)} hrs/día por empleado.
-            </div>
-          )}
-          
-          {/* Conditional text about opportunity vs current */}
-          {roi.current > 0 && roi.opportunity > 0 && (
-            <div className="font-medium text-red-600">
-              El costo de oportunidad (${opportunityMonthlyTotal.toLocaleString()}/mes) es un {Math.round((roi.opportunity / roi.current) * 100)}% del valor que ya generan.
-            </div>
-          )}
-          {/* If current value is zero but there's an opportunity cost */}
-          {roi.current === 0 && roi.opportunity > 0 && (
-            <div className="font-medium text-red-700">
-              Actualmente no se está capitalizando el valor de la IA, perdiendo el potencial de ${opportunityMonthlyTotal.toLocaleString()}/mes.
-            </div>
-          )}
-          <div>(Potencial máx. ahorro: 2 hrs/día/empleado con IA optimizada).</div>
-        </div>
-      </CardContent>
-    </Card>
-  );
-};
-
-const OpportunityCard: React.FC<{ title: string; items: string[]; icon: React.ReactNode; color: string }> = ({
-  title,
-  items,
-  icon,
-  color
-}) => (
-  <Card className="h-full">
-    <CardHeader className="pb-2">
-      <CardTitle className="text-sm font-medium flex items-center gap-2" style={{ color }}>
-        {icon}
-        {title}
-      </CardTitle>
-    </CardHeader>
-    <CardContent className="pt-0">
-      <ul className="space-y-2">
-        {items.map((item, idx) => (
-          <li key={idx} className="flex items-start gap-2 text-sm">
-            <div className="w-1.5 h-1.5 rounded-full mt-2 flex-shrink-0" style={{ backgroundColor: color }} />
-            {item}
-          </li>
-        ))}
-      </ul>
-    </CardContent>
-  </Card>
-);
-
-const SummaryCard: React.FC<{ 
-  title: string; 
-  value: number | string; 
-  icon?: React.ReactNode; 
-  benchmark?: number;
-  unit?: string;
-}> = ({
-  title,
-  value,
-  icon,
-  benchmark,
-  unit = ""
-}) => {
-  // Determine status based on benchmark
-  let statusColor = "#DB4437"; // Red by default
-  let statusText = "Bajo";
-  
-  if (benchmark && typeof value === 'number' && value >= benchmark) {
-    statusColor = "#0F9D58"; // Green
-    statusText = "Excelente";
-  } else if (benchmark && typeof value === 'number' && value >= benchmark * 0.7) {
-    statusColor = "#F4B400"; // Yellow
-    statusText = "Bueno";
-  }
-
-  return (
-    <Card className="flex flex-col justify-between h-full border-l-4" style={{ borderLeftColor: statusColor }}>
-      <CardHeader className="pb-2 flex flex-row items-center justify-between space-y-0">
-        <CardTitle className="text-sm font-medium text-muted-foreground flex items-center gap-2">
-          {icon}
-          {title}
-        </CardTitle>
-        <Badge style={{ backgroundColor: statusColor, color: 'white' }} className="text-xs">
-          {statusText}
-        </Badge>
-      </CardHeader>
-      <CardContent className="pt-0">
-        <div className="flex items-baseline gap-1">
-          <span className="text-3xl font-semibold text-primary">{value}</span>
-          <span className="text-sm text-muted-foreground">{unit}</span>
-        </div>
-        {benchmark && (
-          <div className="mt-2 text-xs text-muted-foreground">
-            Meta: {benchmark}{unit}
-          </div>
-        )}
-      </CardContent>
-    </Card>
-  );
-};
-
-// New component for displaying tasks by area
-interface TasksByAreaProps {
-  data: StatKV[]; // We'll cast the data internally
-}
-
-const TasksByAreaSection: React.FC<TasksByAreaProps> = ({ data }) => {
-  // Extract tasks by area data from the flat data structure
-  const tasksByArea: { [area: string]: { tasks: string[], count: number } } = {};
-  
-  // Parse the flattened data back into structured format
-  data.forEach((item: StatKV) => {
-    if (item.key.startsWith('tasksByArea.') && item.key.endsWith('.tasks')) {
-      const area = item.key.split('.')[1];
-      if (!tasksByArea[area]) {
-        tasksByArea[area] = { tasks: [], count: 0 };
-      }
-      // Cast value to string since tasks are stored as comma-separated strings
-      tasksByArea[area].tasks = item.value ? String(item.value).split(', ') : [];
-    }
-    if (item.key.startsWith('tasksByArea.') && item.key.endsWith('.count')) {
-      const area = item.key.split('.')[1];
-      if (!tasksByArea[area]) {
-        tasksByArea[area] = { tasks: [], count: 0 };
-      }
-      tasksByArea[area].count = Number(item.value) || 0;
-    }
-  });
-
-  // Prepare data for chart
-  const chartData = Object.entries(tasksByArea).map(([area, data]) => ({
-    area: area.charAt(0).toUpperCase() + area.slice(1),
-    count: data.count,
-    tasks: data.tasks
-  })).filter(item => item.count > 0);
-
-  // Area colors mapping
-  const areaColors: { [key: string]: string } = {
-    'Ventas': '#3B82F6',
-    'Marketing': '#10B981', 
-    'Finanzas': '#F59E0B',
-    'Administracion': '#8B5CF6',
-    'Operaciones': '#EF4444',
-    'Recursos humanos': '#06B6D4',
-    'Otro': '#6B7280'
-  };
-
-  if (chartData.length === 0) {
-    return (
-      <div className="text-center py-8 text-gray-500">
-        <p>No se encontraron tareas repetitivas categorizadas para esta empresa.</p>
-      </div>
-    );
-  }
-
-  return (
-    <div className="space-y-6">
-      <div className="mb-4">
-        <h3 className="text-lg font-semibold text-gray-700 mb-2">
-          Distribución de Tareas por Área Funcional
-        </h3>
-        <p className="text-sm text-gray-600">
-          Análisis de las tareas más repetitivas identificadas por área de trabajo
-        </p>
-      </div>
-
-      {/* Chart */}
-      <div className="h-64 w-full">
-        <ResponsiveContainer width="100%" height="100%">
-          <BarChart data={chartData} margin={{ top: 20, right: 30, left: 20, bottom: 5 }}>
-            <CartesianGrid strokeDasharray="3 3" stroke="#f0f0f0" />
-            <XAxis 
-              dataKey="area" 
-              tick={{ fontSize: 12 }}
-              angle={-45}
-              textAnchor="end"
-              height={80}
-            />
-            <YAxis tick={{ fontSize: 12 }} />
-            <Tooltip 
-              content={({ active, payload, label }) => {
-                if (active && payload && payload.length > 0) {
-                  const data = payload[0].payload;
-                  return (
-                    <div className="bg-white p-3 border rounded shadow-lg max-w-xs">
-                      <p className="font-semibold">{label}</p>
-                      <p className="text-sm text-blue-600">
-                        {data.count} tarea{data.count !== 1 ? 's' : ''} identificada{data.count !== 1 ? 's' : ''}
-                      </p>
-                      {data.tasks.length > 0 && (
-                        <div className="mt-2">
-                          <p className="text-xs font-medium text-gray-600">Ejemplos:</p>
-                          <ul className="text-xs text-gray-500 mt-1">
-                            {data.tasks.slice(0, 3).map((task: string, idx: number) => (
-                              <li key={idx} className="truncate">• {task}</li>
-                            ))}
-                            {data.tasks.length > 3 && (
-                              <li className="text-gray-400">... y {data.tasks.length - 3} más</li>
-                            )}
-                          </ul>
-                        </div>
-                      )}
-                    </div>
-                  );
-                }
-                return null;
-              }}
-            />
-            <Bar 
-              dataKey="count" 
-              fill="#3B82F6"
-              radius={[4, 4, 0, 0]}
-            />
-          </BarChart>
-        </ResponsiveContainer>
-      </div>
-
-      {/* Detailed breakdown */}
-      <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-4 mt-6">
-        {chartData.map((areaData, index) => (
-          <Card key={index} className="border-l-4" style={{ borderLeftColor: areaColors[areaData.area] || '#6B7280' }}>
-            <CardHeader className="pb-3">
-              <CardTitle className="text-sm font-medium flex items-center justify-between">
-                <span>{areaData.area}</span>
-                <span className="text-xs bg-blue-100 text-blue-800 px-2 py-1 rounded-full">
-                  {areaData.count} tarea{areaData.count !== 1 ? 's' : ''}
-                </span>
-              </CardTitle>
-            </CardHeader>
-            <CardContent className="pt-0">
-              <div className="space-y-2">
-                {areaData.tasks.slice(0, 3).map((task, idx) => (
-                  <div key={idx} className="text-xs text-gray-600 bg-gray-50 p-2 rounded">
-                    {task}
-                  </div>
-                ))}
-                {areaData.tasks.length > 3 && (
-                  <div className="text-xs text-gray-400 italic">
-                    ... y {areaData.tasks.length - 3} tarea{areaData.tasks.length - 3 !== 1 ? 's' : ''} adicional{areaData.tasks.length - 3 !== 1 ? 'es' : ''}
-                  </div>
-                )}
-              </div>
-            </CardContent>
-          </Card>
-        ))}
-      </div>
-
-      {/* Insights */}
-      <div className="bg-blue-50 p-4 rounded-lg">
-        <h4 className="font-medium text-blue-900 mb-2">💡 Oportunidades de Automatización</h4>
-        <p className="text-sm text-blue-800">
-          Las tareas repetitivas identificadas representan excelentes oportunidades para implementar 
-          soluciones de IA. Las áreas con mayor concentración de tareas podrían beneficiarse de 
-          capacitación especializada en herramientas de automatización.
-        </p>
-      </div>
-    </div>
-  );
 };
 
 const CompanyIADashboard: React.FC<Props> = ({ companyName, stats }) => {
